Refactor: Restructure helper functions and imports
Moved helper methods and object management classes to `helperfunc` module for better organization. Updated all relevant import paths across the project to reflect these changes. Simplified and standardized import structure for improved maintainability.
