+++ src/BeastyBotMain.py
@@ -1,9 +1,9 @@
 import subprocess
-from twitchio.ext import commands, routines
+from twitchio.ext import commands
 
-from config.base_values import CHANNEL_NAME, setup_logger, get_valid_token
-from objects.object_manager import UserManager, ObjectManager
-from objects.setting import BotList
+from src.helperfunc.base_values import CHANNEL_NAME, setup_logger, get_valid_token
+from helperfunc.object_manager import UserManager, ObjectManager
+from helperfunc.setting import BotList
 
 # TODO: Channelpoint integration
