+++ src/obs_bot.py
@@ -1,16 +1,12 @@
 import os
 
-from obsws_python import ReqClient
-from twitchio.ext import commands, routines
+from twitchio.ext import commands
 
-from config.global_methods import get_discussion_topic_for_technology, translate_text, redis_client
-from config.base_values import CHANNEL_NAME, setup_logger, get_valid_token
-from objects.object_manager import UserManager, ObjectManager
+from src.helperfunc.base_values import CHANNEL_NAME, setup_logger, get_valid_token
+from helperfunc.object_manager import UserManager, ObjectManager
 import obsws_python as obs
 
 _logger = setup_logger(__name__)
-
-#TODO: DO STUFF
 
 class Bot(commands.Bot):
 	def __init__(self):
