+++ src/viewer_bot.py
@@ -6,12 +6,12 @@ from typing import TYPE_CHECKING
 import pytz
 from twitchio.ext import commands
 
-from config.global_methods import inc_global_chat_counter, let_ai_narrate_the_fight
-from config.base_values import CHANNEL_NAME, get_valid_token, setup_logger
-from objects.object_manager import ObjectManager
+from src.helperfunc.global_methods import inc_global_chat_counter, let_ai_narrate_the_fight
+from src.helperfunc.base_values import CHANNEL_NAME, get_valid_token, setup_logger
+from helperfunc.object_manager import ObjectManager
 
 if TYPE_CHECKING:
-	from objects.object_manager import UserManager
+	from helperfunc.object_manager import UserManager
 
 _logger = setup_logger(__name__)
